requirements.txt
streamlit==1.44.1
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
python-dateutil==2.8.2
xlsxwriter==3.2.2
# Optional: For interactive charts
plotly==5.18.0